Reorganized the code to extract the plugin from the QRCodeStep file
